Move azure provider pf client test
